Added support for inbound peers. Added node callback for when a connected peer is considered healthy.

+++ common.cpp
@@ -56,7 +56,7 @@ const int Common::BYTES_IN_A_KILOBYTE = 1024;
 const uint16_t Common::HTTP_PORT = 80;
 
 // Tor address length
-const int Common::TOR_ADDRESS_LENGTH = 56;
+const size_t Common::TOR_ADDRESS_LENGTH = 56;
 
 
 // Global variables
